Revert "Load profile data"
