+++ MOBILE/app/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     alias(libs.plugins.android.application)
 }
+import java.util.Properties
 
 def localProps = new Properties()
 def localFile = rootProject.file("local.properties")
@@ -83,8 +84,4 @@ dependencies {
     testImplementation libs.junit
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
-
-    // Lombok
-    compileOnly libs.lombok
-    annotationProcessor libs.lombok
 }
