+++ MOBILE/app/src/main/AndroidManifest.xml
@@ -62,7 +62,7 @@
         <activity
             android:name=".activities.CustomerAccountActivation"
             android:exported="true"
-            android:launchMode="singleTask"
+            android:launchMode="singleTask">
             >
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
