--- MOBILE/app/src/main/java/com/project/mobile/DTO/profile/ImageUploadResponse.java
@@ -1,19 +1,0 @@
-package com.project.mobile.DTO.profile;
-
-public class ImageUploadResponse {
-    private boolean success;
-    private String message;
-    private String imgSrc;
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public String getImgSrc() {
-        return imgSrc;
-    }
-}
