--- MOBILE/app/src/main/java/com/project/mobile/DTO/profile/PasswordChangeRequest.java
@@ -1,19 +1,0 @@
-package com.project.mobile.DTO.profile;
-
-public class PasswordChangeRequest {
-    private String currentPassword;
-    private String newPassword;
-
-    public PasswordChangeRequest(String currentPassword, String newPassword) {
-        this.currentPassword = currentPassword;
-        this.newPassword = newPassword;
-    }
-
-    public String getCurrentPassword() {
-        return currentPassword;
-    }
-
-    public String getNewPassword() {
-        return newPassword;
-    }
-}
