+++ MOBILE/app/src/main/java/com/project/mobile/activities/MainActivity.java
@@ -2,8 +2,6 @@ package com.project.mobile.activities;
 
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.view.View;
-import android.widget.TextView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.annotation.NonNull;
@@ -13,12 +11,13 @@ import androidx.appcompat.app.AppCompatActivity;
 import com.google.android.material.navigation.NavigationView;
 import com.project.mobile.ActiveDrives;
 import com.project.mobile.FragmentTransition;
+import com.project.mobile.History;
 import com.project.mobile.HomeRegistered;
 import com.project.mobile.LiveChat;
 import com.project.mobile.R;
 import com.project.mobile.databinding.ActivityMainBinding;
 import com.project.mobile.fragments.DriverHistoryFragment;
-import com.project.mobile.fragments.profile.ProfileFragment;
+import com.project.mobile.fragments.ProfileFragment;
 import com.project.mobile.fragments.Registered.Rides.BookedRidesFragment;
 import com.project.mobile.viewModels.AuthModel;
