+++ MOBILE/app/src/main/java/com/project/mobile/fragments/Registered/Rides/RideDetailsFragment.java
@@ -48,7 +48,7 @@ import java.util.Locale;
 import java.util.TimeZone;
 import java.util.concurrent.TimeUnit;
 
-public class RideDetailsFragmentActive extends Fragment {
+public class RideDetailsFragment extends Fragment {
 
 
     private static final String TAG = "RideDetailsFragment";
@@ -80,19 +80,19 @@ public class RideDetailsFragmentActive extends Fragment {
 
     private int updateCounter = 0;
 
-    public RideDetailsFragmentActive() {
+    public RideDetailsFragment() {
     }
 
-    public static RideDetailsFragmentActive newInstanceWithId(Long rideId) {
-        RideDetailsFragmentActive fragment = new RideDetailsFragmentActive();
+    public static RideDetailsFragment newInstanceWithId(Long rideId) {
+        RideDetailsFragment fragment = new RideDetailsFragment();
         Bundle args = new Bundle();
         args.putLong(ARG_RIDE_ID, rideId);
         fragment.setArguments(args);
         return fragment;
     }
 
-    public static RideDetailsFragmentActive newInstanceWithAccessToken(String accessToken) {
-        RideDetailsFragmentActive fragment = new RideDetailsFragmentActive();
+    public static RideDetailsFragment newInstanceWithAccessToken(String accessToken) {
+        RideDetailsFragment fragment = new RideDetailsFragment();
         Bundle args = new Bundle();
         args.putString(ARG_ACCESS_TOKEN, accessToken);
         fragment.setArguments(args);
