+++ MOBILE/app/src/main/res/layout/activity_forget_password.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".ForgetPasswordActivity"
-    android:background="@color/background_dark"
+    android:background="@color/background_dark">
     >
     <androidx.appcompat.widget.Toolbar
     android:id="@+id/toolbar"
