+++ MOBILE/app/src/main/res/layout/fragment_vehicle_data.xml
@@ -25,12 +25,12 @@
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:orientation="vertical"
+                android:orientation="horizontal"
                 android:padding="24dp">
 
                 <!-- Vehicle Info Fields -->
                 <LinearLayout
-                    android:layout_width="match_parent"
+                    android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
                     android:orientation="vertical"
@@ -115,7 +115,7 @@
 
                 <!-- Additional Services -->
                 <LinearLayout
-                    android:layout_width="match_parent"
+                    android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
                     android:orientation="vertical">
@@ -145,7 +145,19 @@
 
         </androidx.cardview.widget.CardView>
 
-        <!-- NOTE: btnSaveVehicle has been REMOVED - saving is now done via parent fragment's btnSaveAll -->
+        <!-- Save Button -->
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/btnSaveVehicle"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="16dp"
+            android:text="Save changes"
+            android:textColor="@color/text_dark"
+            android:backgroundTint="@color/accent"
+            app:icon="@drawable/ic_check"
+            app:iconTint="@color/text_dark"
+            app:cornerRadius="8dp"
+            android:padding="14dp"/>
 
     </LinearLayout>
