+++ MOBILE/app/src/main/res/values/strings.xml
@@ -58,5 +58,4 @@
     <string name="destination">Destination:</string>
     <string name="scheduled">Scheduled:</string>
     <string name="cancellation_information">Cancellation Information</string>
-    <string name="profile_save_changes">Save Changes</string>
 </resources>
